+++ share/p2p/shrexsub/pubsub_test.go
@@ -1,4 +1,4 @@
-package shrexpush
+package shrexsub
 
 import (
 	"context"
